Remove bashism from `make check`
The tests run by `make check` require Bash. On Debian you'd need to run
the test as `make SHELL=/bin/bash check`. This commit makes it work on
all POSIX compatible shells (tested on Debian with `dash`).

This is a follow-up on #1225 which was reverted as it didn't create the
logs files correctly.