Encode keys differently
This is the first part of thw work to make the key encoding more flexible.
It will be possible to encode also strings. For now things are hard-coded
to doubles with 2 dimensions.

The C-API is disabled for now, it will come back once the new encoding
is fully in place.