Change the R-tree API
Instead of setting the bounding box on `Seek()`, set it on the
iterator creation. This change was needed as the iterators might
get recreated, hence information passed in by `Seek()` would get
lost. Also `Seek()` is used internally within the Memtable (e.g.
for updated, hence it should behave like it is expected for a
one-dimensional index).